Further adjustments to webtwik (client side rendering using twikifier).
Because of cross site constraints don't use server.host, get the tiddler
data from the current host, blocking ControlView.

It's not clear why CORS is not working here. Probably because of a long
standing bad interaction between TiddlySpace and browsers wherein a 304
response doesn't in include the necessary Allow-Origin header.

Will need to revisit such issues as we move forward.

If a server.host is present it is used, but otherwise, we build the URI
for the collection from the browser's location.