Get command line working for server-side rendering.
twikified.py has been extended to also be wikitext renderer. It uses
the subprocess module to open a pipe to twikify. The current collection
(bag or recipe tiddlers uri) is loaded into twikify, and the text of
the curren tiddler is passed on STDIN. The results come on STDOT and
are returned from the renderer to the serialization.

tiddlywebconfig.py can be used to control whether server side or
client rendering is done.

Problems:

* **Only content readable by GUEST can be rendered.** Short term may
  be able to tweak on the current user such that for HTML we render
  client side and for Atom we render with wikklytext when doing
  private content. Longer term there may be some ways to engage auth.
* twikify does no caching, so when rendering many tiddlers from the
  same collection there can be a lot of reduncancy. The current model
  does not easily allow a long running twikify process, but one could
  be possible. Might better to have an on disk or in memcached cache
  and do ETag matching.
* What should happen for a tiddler that is in search results.
  The only reason for passing a collection is to get macros that look
  at the store working (somewhat) so search results do not a reasonable
  store make. In the current code each tiddler gets its own collection
  based on its recipe or bag. In search results tiddlers only have bags.