A rudimentary unix socket based server that will format a tiddler.
You pass the uri of a collection and the name of a tiddler in that
collection down the socket (space separated)[1] and you will get
back the HTML formatted text.

The process keeps a cache of tiddlers, per collection uri, which means
that subsequent requests for tiddlers from the same collection do
not have the cost of fetching tiddlers.

There is, as yet, no way to dirty the cache but that would not
be difficult to do, either by adding some commands on the socket,
or by switching to using a memcached based cache which is dirtied
by the same tools that twp.cachingstore or twp.etagcache use.

[1] This should be switched to something like \0 separated so we
don't have to worry about escaping in tiddler titles.