when twikifier service is down fall back to raw text
this wraps the raw text of a tiddler in a pre block and renders it
with an error message saying there was an issue rendering the tiddler.

this is for the situation where the twikifier web service is down and
thus cannot render tiddlers properly.

we felt it was better to send something then to throw an error in this
case.