Adjust server.js and twikified to process transclusions more efficiently.
The previous code loaded the collection "fat". This new code
instead passes the text of the current tiddler down the pipe and
that is rendered. The current store is populated skinny, allowing
for most macros.

On the Python side, the output is looking for spans that appear to
be transclusions, finding the desired titles, and rendering them,
and then injecting them into the output.

This should speed up processing on the node side a good deal as
the text of many tiddlers does not need to be messed with...