Handle the socket conversation with proper half shutdowns.
With the change to server.js to accept the tiddler text on the
socket, we need to be able to deal with potentially very long
data strings. To do we must a) shutdown the writing socket halfway,
b) listen for FIN on server.js end in order to ensure we get all the
data. If we don't do this for each 'data' event, a new tiddler is
attempted to be read, which makes the system very confused.

Unfortunately socket half shutdown does not work on on OS X (this is one
of several weirdness with sockets on OS X) which makes it so
server.js no longer works on OS X. This is not a problem because using
twikifier is a high performance optimization.