Remote space transclude from recipe not bag
fix #23

Extract the transclusion section to its own function.

There's a _lot_ more refactoring that could happen here. The
socket handling is very noisy and the dom parsing annoying.

In the proces discover several bugs with the socket handling
in server.js. The essential changes was to return the emitter
from processData before actually starting the function. The
resulting code is not in the correct form, but I'll need to
research the correct form to get it right.