Updates to get twikifier running on the new tiddlyspace.com
The main change here is to make loadRemoteTiddlers return
both an emitter and a function to start the body of the
activity. Without this the function would be done before
anything was ready to hear the events.

The change in twikified.py is make sure that twikifier is
not just the default renderer, but also the renderer for
things which have an explicit tiddlywiki type.