Manage a memory overuse or crash bug
If we properly manage memory by calling window.close (see
commented code in server.js) then the worker may crash. So
instead after maxClientConnections, we gracefully close the
worker's server and exit. The master will automatically start
up another.

In the process of diagnosing this problem other changes were implemented:

* make sure package.json is correct
* submitted a jsdom bug https://github.com/tmpvar/jsdom/issues/614
* fixed a big with argument passing when memcached not being used
* adjusted logging so non-error stuff goes to stdout, and error
  stuff to stderr
* changed some event listeners from 'on' to 'once' or 'addListener'
  to 'on'.

Note the problem at the issue is not yet fixed and until it is
twikifier will be a bit fat, when the problem is figured out
it will use less memory and will not need this restart.