Set a time out on the http request
This helps alleviate 'timeout on socket communication' errors
that are sometimes seen. The path that leads to that are:

* user edits a tiddler with a macro in a large space
* returns to view the rendering of that tiddler
* twikifier loads up the json of the space's tiddlers
* that json is not cached in memcached because of the recent
  change in the space
* the http request goes beyond 10 seconds
* engaging the timeout on the unix socket (between python and node)

This change makes it so the http request can timeout (after 8
seconds) but still allow the tiddler to render (without the
store for the tiddlywiki being filled).

The assumption here is that some rendering is better than a
weird error.

BTW: these timeout errors are _very_ rare, but happen enough
that it seems worth putting some protection in.