Fix #32 to set reasonable classes on tiddlylink v externallinks
getTiddlyLinkInfo is used to retrieve class information for the
title in question. Besides providing a `tiddlyLink` class it also

* labels shadows
* does existence detection and classes as such

This latter is not entirely reliable: The store is only populated
when there are macros in place that would like a store to exist.
This is a handy optimization that fow now we don't want to get rid
of. We'll leave it as a known bug for future discovery.