Use replace task to add create space link at build time
When building twikifier, provide an optional flag to grunt to build the code with a custom create space link function.  This used to be a function hard-coded to tiddlyspace.com

This flag represents the target installation and the default is 'tiddlyspace'

The README will be updated to reflect this when ready.